+++ EveUtils.Migrations.Client.Sqlite/Migrations/20260912011209_AddRunGroupOriginServerAddress.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using EveUtils.Shared.Data;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 #nullable disable
@@ -10,9 +11,11 @@ using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 namespace EveUtils.Migrations.Client.Sqlite.Migrations
 {
     [DbContext(typeof(ClientDbContext))]
-    partial class ClientDbContextModelSnapshot : ModelSnapshot
+    [Migration("20260912011209_AddRunGroupOriginServerAddress")]
+    partial class AddRunGroupOriginServerAddress
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder.HasAnnotation("ProductVersion", "10.0.11");
@@ -1016,42 +1019,6 @@ namespace EveUtils.Migrations.Client.Sqlite.Migrations
                     b.ToTable("RunLootEntry");
                 });
 
-            modelBuilder.Entity("EveUtils.Shared.Modules.Runs.Entities.RunMiningEntry", b =>
-                {
-                    b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd()
-                        .HasColumnType("TEXT");
-
-                    b.Property<int>("CriticalUnits")
-                        .HasColumnType("INTEGER");
-
-                    b.Property<DateTime>("FirstObservedAtUtc")
-                        .HasColumnType("TEXT");
-
-                    b.Property<DateTime>("LastObservedAtUtc")
-                        .HasColumnType("TEXT");
-
-                    b.Property<string>("OreType")
-                        .IsRequired()
-                        .HasColumnType("TEXT");
-
-                    b.Property<int>("ResidueUnits")
-                        .HasColumnType("INTEGER");
-
-                    b.Property<Guid>("RunId")
-                        .HasColumnType("TEXT");
-
-                    b.Property<int>("Units")
-                        .HasColumnType("INTEGER");
-
-                    b.HasKey("Id");
-
-                    b.HasIndex("RunId", "OreType")
-                        .IsUnique();
-
-                    b.ToTable("RunMiningEntry");
-                });
-
             modelBuilder.Entity("EveUtils.Shared.Modules.Runs.Entities.RunParameter", b =>
                 {
                     b.Property<Guid>("Id")
@@ -1466,17 +1433,6 @@ namespace EveUtils.Migrations.Client.Sqlite.Migrations
                     b.Navigation("RunLootCapture");
                 });
 
-            modelBuilder.Entity("EveUtils.Shared.Modules.Runs.Entities.RunMiningEntry", b =>
-                {
-                    b.HasOne("EveUtils.Shared.Modules.Runs.Entities.Run", "Run")
-                        .WithMany("MiningEntries")
-                        .HasForeignKey("RunId")
-                        .OnDelete(DeleteBehavior.Cascade)
-                        .IsRequired();
-
-                    b.Navigation("Run");
-                });
-
             modelBuilder.Entity("EveUtils.Shared.Modules.Runs.Entities.RunParameter", b =>
                 {
                     b.HasOne("EveUtils.Shared.Modules.Runs.Entities.Run", "Run")
@@ -1506,8 +1462,6 @@ namespace EveUtils.Migrations.Client.Sqlite.Migrations
                     b.Navigation("EnemyObservations");
 
                     b.Navigation("LootCaptures");
-
-                    b.Navigation("MiningEntries");
 
                     b.Navigation("Parameters");
                 });